+++ lib/OAuth2ResponseHandler.php
@@ -23,11 +23,13 @@ class OAuth2ResponseHandler
      * Look at the state parameter returned by the OAuth2 server and determine if we can handle it;
      * @return bool true if response can be handled by this module
      */
-    public function canHandleResponse() {
+    public function canHandleResponse()
+    {
         return $this->canHandleResponseFromRequest($_REQUEST);
     }
 
-    public function canHandleResponseFromRequest(array $request) {
+    public function canHandleResponseFromRequest(array $request)
+    {
         return strpos(@$request['state'], $this->expectedPrefix) === 0;
     }
 
@@ -38,17 +40,20 @@ class OAuth2ResponseHandler
     {
         $this->handleResponseFromRequest($_REQUEST);
     }
-    public function handleResponseFromRequest(array $request) {
+
+    public function handleResponseFromRequest(array $request)
+    {
         //TODO: use some error checking
         Logger::debug('authoauth2 : linkback request=' . var_export($request, true));
 
         if (!$this->canHandleResponseFromRequest($request)) {
+            // phpcs:ignore Generic.Files.LineLength.TooLong
             throw new \SimpleSAML_Error_BadRequest('Either missing state parameter on OAuth2 login callback, or cannot be handled by authoauth2');
         }
         $stateIdWithPrefix = $request['state'];
         $stateId = substr($stateIdWithPrefix, strlen($this->expectedPrefix));
-        //TODO: decide how no-state errors should be handled? Likely cause is user clicked back button (state was already consumed and removed)
-        // or session expired
+        //TODO: decide how no-state errors should be handled?
+        // Likely cause is user clicked back button (state was already consumed and removed) or session expired
         $state = \SimpleSAML_Auth_State::loadState($stateId, $this->expectedStageState);
 
         // Find authentication source
@@ -61,7 +66,7 @@ class OAuth2ResponseHandler
          * @var OAuth2 $source
          */
         $source = \SimpleSAML_Auth_Source::getById($sourceId, OAuth2::class);
-        if ($source === NULL) {
+        if ($source === null) {
             throw new \SimpleSAML_Error_BadRequest('Could not find authentication source with id ' . $sourceId);
         }
         if (!array_key_exists('code', $request)) {
@@ -71,17 +76,25 @@ class OAuth2ResponseHandler
         try {
             $source->finalStep($state, $request['code']);
         } catch (IdentityProviderException $e) {
+            // phpcs:ignore Generic.Files.LineLength.TooLong
             Logger::error("authoauth2: error in '$sourceId' msg '{$e->getMessage()}' body '" . var_export($e->getResponseBody(), true) . "'");
-            \SimpleSAML_Auth_State::throwException($state, new \SimpleSAML_Error_AuthSource($sourceId, 'Error on oauth2 linkback endpoint.', $e));
+            \SimpleSAML_Auth_State::throwException(
+                $state,
+                new \SimpleSAML_Error_AuthSource($sourceId, 'Error on oauth2 linkback endpoint.', $e)
+            );
         } catch (\Exception $e) {
             Logger::error("authoauth2: error in '$sourceId' msg '{$e->getMessage()}'");
-            \SimpleSAML_Auth_State::throwException($state, new \SimpleSAML_Error_AuthSource($sourceId, 'Error on oauth2 linkback endpoint.', $e));
+            \SimpleSAML_Auth_State::throwException(
+                $state,
+                new \SimpleSAML_Error_AuthSource($sourceId, 'Error on oauth2 linkback endpoint.', $e)
+            );
         }
 
         \SimpleSAML_Auth_Source::completeAuth($state);
     }
 
-    private function handleErrorResponse(\SimpleSAML_Auth_Source $source,  array $state, array $request) {
+    private function handleErrorResponse(\SimpleSAML_Auth_Source $source, array $state, array $request)
+    {
         // Errors can be pretty inconsistent
         $error = @$request['error'];
         // 'access_denied' is OAuth2 standard. Some AS made up their own codes, so support the common ones.
@@ -91,11 +104,9 @@ class OAuth2ResponseHandler
             \SimpleSAML_Auth_State::throwException($state, $e);
         }
 
-        $errorMsg = 'Authentication failed: ['. $error.'] '. @$request['error_description'];
+        $errorMsg = 'Authentication failed: [' . $error . '] ' . @$request['error_description'];
         Logger::debug("Authsource '" . $source->getAuthId() . "' return error $errorMsg");
-        $e = new \SimpleSAML_Error_AuthSource($source->getAuthId(), $errorMsg );
+        $e = new \SimpleSAML_Error_AuthSource($source->getAuthId(), $errorMsg);
         \SimpleSAML_Auth_State::throwException($state, $e);
-
     }
-
 }