+++ CHANGELOG.md
@@ -4,6 +4,11 @@
 
 _Release: 2018-?
 
+* Use ssp 1.16.2 as the dependency
+* Add php 7.1 and 7.2 to travis buildds
+* PSR-2 styling
+* Add Microsoft authsource
+* Allow logging of id_token json
 * Template for YahooOIDC, MicrosoftOIDC, LinkedIn and Facebook
 * Add support for enabling http request/response logging
 * Add general debug information