+++ CHANGELOG.md
@@ -3,7 +3,10 @@
 ## Unreleased
 
 _Release: 2018-?
-
+* Behavior changes from v1
+    * User canceling consent sends them to error page rather than throwing USER_ABORT. Behavior is configurable
+    * Automatic retry on network errors. Behavior is configurable
+* If user cancels consent, send them to page saying consent must be provided.
 * Perform 1 retry on network errors
 * Use ssp 1.16.2 as the dependency
 * Add php 7.1 and 7.2 to travis buildds